VK API limitations fixes
- WARN: app now have vulnerability

Due to limitations in access_token that we get from frontend and take on
backend -- issue appeared in production. We cannot send api requests
from IP different from where access_token was received.

There was a way to workaround as I see:

- create another app (website type)
- use it only to obtain token on server side

Cons of this approach is that application complexity tremendously grows,
and user will have to pass through TWO authorizations on the same
website what looks suspiciously

For now I'll keep it as is, beacuse it's training project

Changes:

- removed vk calls from serverside and moved to frontend